Deleted Progress Comp/Copied Code To The Main Component
